Set RPATH for installed targets and fix Qt deployment on Linux
Set CMAKE_INSTALL_RPATH to $ORIGIN for executables, modules, and
plugins. Fix SuperBuild external project RPATH to include $ORIGIN.
Add RPATH for autoload modules and CTK plugins pointing to the
parent bin/ directory. Override QT_DEPLOY_* variables to redirect
Qt libraries into bin/ on Linux. Generate qt.conf with relative
prefix. Use build-tree binary for Qt plugin detection to support
Qt 6.10's install-time module resolution. Simplify wrapper scripts
to only set LD_LIBRARY_PATH for Python. Remove GCC plugin preload
workaround and Qt environment variables from applications.

+++ CMake/mitkFunctionDeployQt.cmake
@@ -27,12 +27,30 @@ function(mitkFunctionDeployQt _target)
         string(APPEND _deploy_tool_options " \"--openssl-root\" \"${_openssl_root}\"")
       endif()
     endif()
+
+    # On Linux, override QT_DEPLOY_* variables so that everything (including
+    # the WebEngine deployment hook) goes under bin/ instead of the default
+    # lib/, libexec/, plugins/, translations/ at the prefix root.
+    set(_deploy_prefix_overrides "")
+    if(LINUX)
+      set(_deploy_prefix_overrides "
+set(QT_DEPLOY_BIN_DIR \"bin\")
+set(QT_DEPLOY_LIB_DIR \"bin\")
+set(QT_DEPLOY_LIBEXEC_DIR \"bin\")
+set(QT_DEPLOY_PLUGINS_DIR \"bin/plugins\")
+set(QT_DEPLOY_QML_DIR \"bin/qml\")
+set(QT_DEPLOY_TRANSLATIONS_DIR \"bin/translations\")
+set(QT_DEPLOY_DATA_DIR \"bin\")
+")
+    endif()
+
     qt_generate_deploy_script(
       TARGET ${_target}
       OUTPUT_SCRIPT _deploy_script
       CONTENT "
+${_deploy_prefix_overrides}
 qt_deploy_runtime_dependencies(
-  EXECUTABLE \"bin/$<TARGET_FILE_NAME:${_target}>\"
+  EXECUTABLE \"$<TARGET_FILE:${_target}>\"
   BIN_DIR \"bin\"
   LIB_DIR \"bin\"
   PLUGINS_DIR \"bin/plugins\"
@@ -46,58 +64,37 @@ qt_deploy_runtime_dependencies(
   endif()
   install(SCRIPT ${_deploy_script})
 
-  # windeployqt deploys some files (WebEngine resources, translations)
-  # relative to the install prefix root rather than bin/. Move them
-  # into bin/ to keep everything in one directory and remove the
-  # root-level duplicates.
-  install(CODE "
-    foreach(_qt_dir qml resources translations)
-      set(_root_dir \"\${CMAKE_INSTALL_PREFIX}/\${_qt_dir}\")
-      if(IS_DIRECTORY \"\${_root_dir}\")
-        message(STATUS \"MITK_DEPLOY: Moving root-level \${_qt_dir}/ to bin/\${_qt_dir}/\")
-        file(MAKE_DIRECTORY \"\${CMAKE_INSTALL_PREFIX}/bin/\${_qt_dir}\")
-        file(GLOB _items \"\${_root_dir}/*\")
-        foreach(_item \${_items})
-          get_filename_component(_name \"\${_item}\" NAME)
-          set(_dest \"\${CMAKE_INSTALL_PREFIX}/bin/\${_qt_dir}/\${_name}\")
-          if(NOT EXISTS \"\${_dest}\")
-            file(RENAME \"\${_item}\" \"\${_dest}\")
-          endif()
-        endforeach()
-        file(REMOVE_RECURSE \"\${_root_dir}\")
-      endif()
-    endforeach()
-  ")
-
-  # On Linux, qt_deploy_runtime_dependencies() only traces shared library
-  # dependencies. QtWebEngine additionally requires a helper executable,
-  # data files, and locale files that must be deployed manually.
-  if(LINUX AND TARGET Qt6::WebEngineCore)
+  # On Linux the WebEngine deployment hook generates qt.conf with an
+  # absolute staging path as prefix. Overwrite it with a correct
+  # relative prefix after the deploy script has run.
+  if(LINUX)
     install(CODE "
-      set(_webengine_process \"${QT_INSTALL_LIBEXECS}/QtWebEngineProcess\")
-      if(EXISTS \"\${_webengine_process}\")
-        message(STATUS \"MITK_DEPLOY: Installing QtWebEngineProcess\")
-        file(INSTALL \"\${_webengine_process}\"
-          DESTINATION \"\${CMAKE_INSTALL_PREFIX}/bin\"
-          FILE_PERMISSIONS OWNER_READ OWNER_WRITE OWNER_EXECUTE GROUP_READ GROUP_EXECUTE WORLD_READ WORLD_EXECUTE
-        )
-      endif()
+      set(_qt_conf \"\${CMAKE_INSTALL_PREFIX}/bin/qt.conf\")
+      message(STATUS \"MITK_DEPLOY: Writing \${_qt_conf}\")
+      file(WRITE \"\${_qt_conf}\" \"[Paths]\nPrefix = .\n\")
+    ")
+  endif()
 
-      set(_webengine_resources \"${QT_INSTALL_DATA}/resources\")
-      if(IS_DIRECTORY \"\${_webengine_resources}\")
-        message(STATUS \"MITK_DEPLOY: Installing QtWebEngine resources\")
-        file(INSTALL \"\${_webengine_resources}/\"
-          DESTINATION \"\${CMAKE_INSTALL_PREFIX}/bin/resources\"
-        )
-      endif()
-
-      set(_webengine_locales \"${QT_INSTALL_TRANSLATIONS}/qtwebengine_locales\")
-      if(IS_DIRECTORY \"\${_webengine_locales}\")
-        message(STATUS \"MITK_DEPLOY: Installing QtWebEngine locale en-US.pak\")
-        file(INSTALL \"\${_webengine_locales}/en-US.pak\"
-          DESTINATION \"\${CMAKE_INSTALL_PREFIX}/bin/translations/qtwebengine_locales\"
-        )
-      endif()
+  if(WIN32)
+    # windeployqt places resources and translations at the prefix root.
+    # Move them into bin/ to keep everything in one directory.
+    install(CODE "
+      foreach(_qt_dir resources translations)
+        set(_root_dir \"\${CMAKE_INSTALL_PREFIX}/\${_qt_dir}\")
+        if(IS_DIRECTORY \"\${_root_dir}\")
+          message(STATUS \"MITK_DEPLOY: Moving root-level \${_qt_dir}/ to bin/\${_qt_dir}/\")
+          file(MAKE_DIRECTORY \"\${CMAKE_INSTALL_PREFIX}/bin/\${_qt_dir}\")
+          file(GLOB _items \"\${_root_dir}/*\")
+          foreach(_item \${_items})
+            get_filename_component(_name \"\${_item}\" NAME)
+            set(_dest \"\${CMAKE_INSTALL_PREFIX}/bin/\${_qt_dir}/\${_name}\")
+            if(NOT EXISTS \"\${_dest}\")
+              file(RENAME \"\${_item}\" \"\${_dest}\")
+            endif()
+          endforeach()
+          file(REMOVE_RECURSE \"\${_root_dir}\")
+        endif()
+      endforeach()
     ")
   endif()
 endfunction()
